updated tests to see page changes, moved linkButton to its ownb file
