+++ frontend/package.json
@@ -36,9 +36,9 @@
     "react-accessible-dropdown-menu-hook": "^3.1.0",
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.15.0",
+    "react-responsive": "^9.0.0-beta.4",
     "topojson-client": "^3.1.0",
-    "topojson-simplify": "^3.0.3",
-    "react-responsive": "^9.0.0-beta.4"
+    "topojson-simplify": "^3.0.3"
   },
   "devDependencies": {
     "@babel/core": "^7.15.5",
@@ -91,7 +91,7 @@
     "msw-storybook-addon": "^1.3.0",
     "next-router-mock": "^0.3.0",
     "node-fetch": "^3.0.0",
-    "prettier": "2.3.0",
+    "prettier": "^2.3.0",
     "react-table": "^7.7.0",
     "react-test-renderer": "^17.0.2",
     "resize-observer-polyfill": "^1.5.1",
