+++ frontend/pages/forgot/index.tsx
@@ -19,18 +19,16 @@ export default function ForgotPassword() {
   const form = useForm()
   const [loading, setLoading] = useState(false)
   const [submitError, setSubmitError] = useState(null)
-  const [useEmail, setEmailState] = useState(true) //true when using email, false when using phone number
+  const [useEmail, setEmailState] = useState(true)
 
   async function onSubmit(formValues: any) {
     setLoading(true)
     setSubmitError(null)
     try {
       if(useEmail){
-        //find password using email
         setSubmitError("Email not connected to back end")
       }
       else{
-        //find password using phone number
         setSubmitError("Phone number not connected to back end")
       }
     } catch (e) {
